Move drawing of ledger beam to its own class and start cleanup of drawing code.
