+++ CadRevealFbxProvider.Tests/BatchUtils/ScaffoldOptimizer/GeometryInstancerTests.cs
